server/apps/chatbot-service/requirements.txt
kafka-python==2.1.1
redis==3.5.3
google-generativeai==0.8.4
google-api-python-client==2.164.0
python-dotenv==1.0.1
openai==1.70.0
aiohttp==3.9.1
